requirements.txt
asgiref==3.2.3
Django==3.0.3
pytz==2019.3
sqlparse==0.3.0
djangorestframework==3.11.0
python-dotenv==0.10.5
psycopg2==2.8.4
